+++ R/comapre.R
@@ -5,7 +5,7 @@
 #' @param new character a new version of package, default newest version. Default: NULL
 #' @param fields character a vector with possible values `c("Depends", "Imports", "LinkingTo", "Suggests")`. Default: `c("Depends", "Imports", "LinkingTo")`
 #' @param lib.loc character. Default: NULL
-#' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
+#' @param repos character the base URL of the repository to use. Used only for the validation. Default `https://cran.rstudio.com/`
 #' @return data.frame with 4 columns.
 #' \describe{
 #' \item{Package}{character package names.}
@@ -25,6 +25,7 @@ pac_compare_versions <- function(pac,
                                  lib.loc = NULL,
                                  repos = "https://cran.rstudio.com/") {
   stopifnot((length(pac) == 1) && is.character(pac))
+  stopifnot(pac %in% rownames(available_packages(repos = repos)))
   stopifnot(is.null(old) || (length(old) == 1) && is.character(old))
   stopifnot(is.null(new) || (length(new) == 1) && is.character(new))
   stopifnot(all(fields %in% c("Depends", "Imports", "Suggests", "LinkingTo")))
@@ -40,18 +41,20 @@ pac_compare_versions <- function(pac,
     new <- pac_last(pac)
   }
 
-  if (utils::compareVersion(new, old) < 0) {
-    return(data.frame())
-  }
+  stopifnot(utils::compareVersion(new, old) >= 0)
 
-  one_base <- paste(Filter(function(x) length(x) > 0, pac_description(pac, version = old, lib.loc = lib.loc, repos = repos)[fields]), collapse = ",")
+  one_desc <- pac_description(pac, version = old, lib.loc = lib.loc, repos = repos)
+  if (length(one_desc) == 0) stop(sprintf("Version %s is not exists for %s.", old, pac))
+  one_base <- paste(Filter(function(x) length(x) > 0, one_desc[fields]), collapse = ",")
   one_e <- extract_deps(one_base)
   s_remote <- unique(data.frame(
     Package = one_e$packages[[1]], Version = replaceNA(one_e$versions[[1]], ""),
     stringsAsFactors = FALSE
   ))
 
-  two_base <- paste(Filter(function(x) length(x) > 0, pac_description(pac, version = new, lib.loc = lib.loc, repos = repos)[fields]), collapse = ",")
+  two_desc <- pac_description(pac, version = new, lib.loc = lib.loc, repos = repos)
+  if (length(two_desc) == 0) stop(sprintf("Version %s is not exists for %s.", new, pac))
+  two_base <- paste(Filter(function(x) length(x) > 0, two_desc[fields]), collapse = ",")
   two_e <- extract_deps(two_base)
   s_remote2 <- unique(data.frame(
     Package = two_e$packages[[1]], Version = replaceNA(two_e$versions[[1]], ""),
@@ -63,6 +66,7 @@ pac_compare_versions <- function(pac,
   col_new <- paste0("Version.", new)
   res$version_status <- apply(res, 1, function(x) utils::compareVersion(x[col_new], x[col_old]))
   rownames(res) <- NULL
+  attr(res, "package") <- pac
   attr(res, "old") <- old
   attr(res, "new") <- new
   res
@@ -74,13 +78,11 @@ pac_compare_versions <- function(pac,
 #' @param old character an old version of package.
 #' @param new character a new version of package.
 #' @param lib.loc character. Default: NULL
-#' @param repos character the base URL of the repositories to use. Default `https://cran.rstudio.com/`
+#' @param repos character the base URL of the repository to use. Used only for the validation. Default `https://cran.rstudio.com/`
 #' @return list with `c("imports", "exports", "exportPatterns", "importClasses", "importMethods", "exportClasses", "exportMethods", "exportClassPatterns", "dynlibs", "S3methods")` slots, and added and removed ones for each of them.
 #' @note The comparison is only about exports.
 #' @export
 #' @examples
-#' pac_compare_namespace("memoise", "0.2.1", "2.0.0")
-#' pac_compare_namespace("memoise", "0.2.1", "2.0.0")$exports
 #' pac_compare_namespace("shiny", "1.0.0", "1.6.0")
 #' pac_compare_namespace("shiny", "1.0.0", "1.6.0")$exports
 pac_compare_namespace <- function(pac,
@@ -89,6 +91,7 @@ pac_compare_namespace <- function(pac,
                                 lib.loc = NULL,
                                 repos = "https://cran.rstudio.com/") {
   stopifnot((length(pac) == 1) && is.character(pac))
+  stopifnot(pac %in% rownames(available_packages(repos = repos)))
   stopifnot(is.null(old) || (length(old) == 1) && is.character(old))
   stopifnot(is.null(new) || (length(new) == 1) && is.character(new))
   stopifnot(is.character(repos))
@@ -103,14 +106,19 @@ pac_compare_namespace <- function(pac,
     new <- pac_last(pac)
   }
 
-  if (utils::compareVersion(new, old) < 0) {
-    return(list())
-  }
+  stopifnot(utils::compareVersion(new, old) >= 0)
+
   result <- list()
   fields <- c("imports", "exports", "exportPatterns", "importClasses", "importMethods", "exportClasses", "exportMethods", "exportClassPatterns", "dynlibs", "S3methods")
+
+  one_nam <- pac_namespace(pac, old, lib.loc = lib.loc, repos = repos)
+  if (length(one_nam) == 0) stop(sprintf("Version %s is not exists for %s.", old, pac))
+  two_nam <- pac_namespace(pac, new, lib.loc = lib.loc, repos = repos)
+  if (length(two_nam) == 0) stop(sprintf("Version %s is not exists for %s.", new, pac))
+
   for (f in fields) {
-    old_f <- unlist(pac_namespace(pac, old, lib.loc = lib.loc, repos = repos)[[f]])
-    new_f <- unlist(pac_namespace(pac, new, lib.loc = lib.loc, repos = repos)[[f]])
+    old_f <- unlist(one_nam[[f]])
+    new_f <- unlist(two_nam[[f]])
 
     result[[f]] <- list(removed = setdiff(old_f, new_f), added = setdiff(new_f, old_f))
   }