+++ R/pacs.R
@@ -1,5 +1,5 @@
 #' @importFrom memoise memoise
 #' @importFrom stringi stri_match_all stri_split_fixed stri_split_regex stri_trim
-#' @importFrom xml2 xml_find_all read_html xml_text
+#' @importFrom xml2 xml_find_all read_html xml_text xml_attr
 #'
 NULL